src/requirements.txt
google-cloud-logging
google-cloud-bigquery
firebase-admin
click
